+++ examples/ui-prompting-examples/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ux-private/ui-prompting-examples
 
+## 0.5.0
+
+### Minor Changes
+
+- 68b0e54: Add Rich Text Editor Button Groups support
+
+### Patch Changes
+
+- Updated dependencies [68b0e54]
+    - @sap-ux/ui-prompting@0.6.0
+
 ## 0.4.14
 
 ### Patch Changes